+++ CHANGELOG.md
@@ -10,6 +10,32 @@ Neste arquivo (CHANGELOG.MD) você encontrará somente as mudanças referentes a
 ### Sobre os "BREAKING CHANGES"
 Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de serem pequenas, podem alterar o comportamento da funcionalidade caso não seja feita uma atualização, **preste muita atenção** nas breaking changes dentro das versões quando existirem.
 
+## Não publicado
+## BREAKING CHANGES
+- `QasNestedFields`: agora propriedade `actions-menu-props` só funciona junto da prop `:use-inline-actions="true"`.
+
+### Adicionado
+- `QasNestedFields`:
+  - adicionado componente `QasHeader` com a prop `header-props`.
+  - adicionado nova propriedade `form-common-columns`.
+
+### Corrigido
+- `QasActions`: corrigido style que não deixava div 100% quando usado com a prop `useFullWidth`.
+- `QasNestedFields`: corrigido problemas de espaçamentos.
+- `QasFormGenerator`: corrigido div vazia quando não tinha fieldset.
+- `QasHeader`: corrigido problemas de layout.
+
+### Modificado
+- `QasDialog`:
+  - alterado espaçamento interno de `lg` para `md`.
+  - alterado titulo de h5 para h4 com componente `QasLabel`.
+  - alterado espaçamento bottom do título de `lg` para `md`.
+- [`QasFormGenerator`, `QasGridGenerator`]: alterado espaçamento gutter de `lg` para `md`.
+- `QasNestedFields`:
+  - modificado set scroll ao adicionado novas linhas, agora o scroll fica fixo na ultima linha adicionada.
+  - agora propriedade `actions-menu-props` só funciona junto da prop `:use-inline-actions="true"`.
+  - default do `formGutter` alterado de `lg` para `md`.
+
 ## [3.17.0-beta.4] - 02-09-2024
 ### Adicionado
 - `QasSelect`: